refactor: types into types last will file
