+++ frontend/src/components/Modal/HeirsModal/HeirsOrganisationModal/HeirsOrganisationModal.tsx
@@ -3,20 +3,11 @@ import { Form, Formik } from 'formik'
 import { ObjectSchema, object, string } from 'yup'
 import { useAppDispatch, useAppSelector } from '../../../../store/hooks'
 import { addOrganisationHeir, sendLastWillState, updateOrganisationHeir } from '../../../../store/lastwill'
-import { Organisation } from '../../../../types/lastWill'
+import { Organisation, OrganisationFormPayload } from '../../../../types/lastWill'
 import { Button } from '../../../ButtonsAndLinks/Button/Button'
 import { TextInput } from '../../../Form/TextInput/TextInput'
 import { Headline } from '../../../Headline/Headline'
 import { Modal } from '../../ModalBase/Modal'
-
-export type OrganisationFormPayload = {
-	id: string
-	name?: string
-	street?: string
-	houseNumber?: string
-	zipCode?: string
-	city?: string
-}
 
 type HeirsOrganisationModalProps = {
 	/** Modal Open/Close State. */
